requirements.txt
torch
torchaudio
torchvision
opencv-python
tqdm
pycocotools
black
Pillow
numpy
